fix tabs, move and resize many windows, add full diseqc 1.1 support ( uncommitted switches ), fix support for Rotors behind diseqc switches, combobox and listbox fixes, improved help support, add scrollbar to help window, add other dupe tsid/onid combinations, fix warings, fix pixmap positions when label have a shortcut pixmap, show streaminfo on doubleklick the info key / ? key, fix serviceselector path destroying... in hope, show messagebox after end of recording... ( show recorded movied?), disable showBouquet action.. ( destroying ssel pathes ) add Simple DiSEqC Option ( many satellites via DiSEqC ), fix DiSEqC Configuration Wizard for Rotor Support

+++ tuxbox/enigma/lib/dvb/si.cpp
@@ -140,7 +140,7 @@ UnknownDescriptor::~UnknownDescriptor()
 eString UnknownDescriptor::toString()
 {
 	std::stringstream res;
-	res << "UnknownDescriptor: " << decode_descr(data[0]) << " (" << std::hex << std::setfill('0') << std::setw(4) << data[0] << ")\n" << "	rawData:";
+	res << "UnknownDescriptor: " << decode_descr(data[0]) << " (" << std::hex << std::setfill('0') << std::setw(4) << (int)data[0] << ")\n" << "	rawData:";
 	for (int i=0; i<len; i++)
 		res << std::setw(2) << data[i];
 	res << std::endl;
